Refactored AssessmentSummaryBuilder
Moved default method in AssessmentOrchestrator to AssessmentSummaryBuilder and renamed to AssessmentSummaryService
